backend/requirements.txt
fastapi==0.68.2
uvicorn[standard]==0.15.0
mysql-connector-python==8.0.33
pydantic==1.10.2
python-multipart==0.0.5
python-dotenv==0.19.2

